+++ dynamic_libs/FloatingRobot.hpp
@@ -22,13 +22,13 @@
 
 class FloatingRobot : public Enemy {
 private:
-    const Engine::Size<int> _size = {SIZE_X, SIZE_Y};
+    const Engine::Size<float> _size = {SIZE_X, SIZE_Y};
 
 public:
-    explicit FloatingRobot(const Engine::Point<int> &pos = {0, 0}) : Enemy(std::move(std::make_unique<DataSprite>(PATH, Engine::Box<int>{pos, {SIZE_X, SIZE_Y}})), pos) {
+    explicit FloatingRobot(const Engine::Point<float> &pos = {0, 0}) : Enemy(std::move(std::make_unique<DataSprite>(PATH, Engine::Box<float>{pos, {SIZE_X, SIZE_Y}})), pos) {
         this->addComponent<PatternComponent>();
         this->addComponent<AutomaticWeaponComponent>(1, 0.5, 0.2);  // FIXME changer les stats
-        this->addComponent<Engine::AnimationComponent>(0.2, std::map<int, std::vector<Engine::Box<int>>>{
+        this->addComponent<Engine::AnimationComponent>(0.2, std::map<int, std::vector<Engine::Box<float>>>{
                 {DEFAULT, {
                     {_size.x * 0, _size.x * 1, 0, _size.y},
                     {_size.x * 1, _size.x * 2, 0, _size.y},
